add test for enum deserialization conflict case
- also consolidate all mock enums into two files
