+++ APIMatic.Core.Test/MockTypes/Models/NumberEnums.cs
